[#10] refactor: 패키지 이동
- ProductContainer(+ 테스트): service => component로 이동
- Scraping(+ 테스트): service => domain으로 이동
